refactor(utils): split file.rs into cohesive submodules
Decompose the oversized `src/utils/file.rs` grab-bag into a directory
module `src/utils/file/` split by concern, re-exporting the public API
from `mod.rs` so all existing `crate::utils::file::*` paths still resolve.

New layout:
- path.rs: filesystem metadata, glob exclusion, extension/name predicates
- encoding.rs: byte-to-text decoding and "looks like text" heuristics
- format_sniff.rs: magic-byte and container-format detection
- classify.rs: mime type and file-info classification surface
- extract.rs: text extraction for license/copyright detection
- image_metadata.rs: EXIF/XMP image metadata text extraction
- pdf.rs: bounded PDF text extraction

Pure refactor: no behavior or golden-fixture changes. Cross-module
internals use `pub(super)` and the public API keeps its `pub`/`pub(crate)`
visibility. In-module unit tests moved next to the code they cover; the
cross-cutting integration tests for the two public entry points stay in
`mod.rs`.

Remediated the single `.unwrap()` in `get_creation_date`: the pre-epoch
`duration_since(UNIX_EPOCH)` failure is now propagated as `None` via `?`
instead of panicking, matching the function's existing `Option` contract.
